+++ linera-views/src/simple_store.rs
